refactor: login 대/소문자 변경하기 위해 Comp로 변경
